transform: move propagatable impl in it's own file
