+++ crates/bevy_transform/src/propagate.rs
@@ -1,7 +1,31 @@
+use bevy_hierarchy::Propagatable;
+
+use crate::{GlobalTransform, Transform};
+
+impl Propagatable for Transform {
+    type Computed = GlobalTransform;
+    type Payload = GlobalTransform;
+
+    const ALWAYS_PROPAGATE: bool = false;
+
+    #[inline]
+    fn compute_root(computed: &mut Self::Computed, local: &Self) {
+        *computed = GlobalTransform::from(*local);
+    }
+
+    #[inline]
+    fn compute(computed: &mut Self::Computed, payload: &Self::Payload, local: &Self) {
+        *computed = payload.mul_transform(*local);
+    }
+
+    #[inline]
+    fn payload(computed: &Self::Computed) -> Self::Payload {
+        *computed
+    }
+}
+
 #[cfg(test)]
 mod test {
-    // TODO: move these tests to bevy_hierarchy::Propagate module
-
     use bevy_app::prelude::*;
     use bevy_ecs::prelude::*;
     use bevy_ecs::system::CommandQueue;
